Fixed potentially suboptimal industry allocation when terraforming. + minor improvements
- Fixed potentially suboptimal industry allocation when terraforming.
- DNA Workshop: removed annoying beep.
  - (It was a false positive error detection, when updating the file list; as for a moment the selection was set to "null")
- Tech Panel: Improved display of total research by removing rounding errors.
- Improved display of "Production per factory". (Now, up to two decimal places.)

+++ src/rotp/ui/tech/AllocateTechUI.java
@@ -517,8 +517,8 @@ public final class AllocateTechUI extends BasePanel implements MouseListener, Mo
         }
 
         TechTree tree = player().tech();
-        int totalSpending = (int) totalPlanetaryResearchSpending;
-        int totalResearch = 0;       
+		float totalSpending = totalPlanetaryResearchSpending;
+		float totalResearch = 0;
         for (int i=0;i<cats;i++)
              totalResearch += tree.category(i).currentResearch(totalPlanetaryResearch());
